+++ lib/app/presentation/pages/item/item_edit_page.dart
@@ -13,7 +13,7 @@ import '../../../application/model/item/selected_image_model.dart';
 import '../../../application/state/locale_provider.dart';
 import '../../../application/usecase/item/item_usecase.dart';
 import '../../../application/usecase/item/state/item_page_providers.dart';
-import '../../../application/usecase/purchase/state/wanter_name_suggestion.dart';
+import '../../../application/usecase/item/state/wanter_name_suggestion.dart';
 import '../../components/importer.dart';
 import '../../routes/src/routes_data.dart';
 import '../error/components/error_view.dart';
